snapshot: in-progress multichannel/build_id work (Electron-port rebase base)
Captures the uncommitted src-tauri + frontend work so the Electron sidecar
port targets the real current surface instead of the older committed HEAD.

+++ src/features/settings/state/useRouteRestore.ts
@@ -68,7 +68,7 @@ export function useRouteRestore() {
         setup_complete: state.settings.setup_complete,
         last_active_route: location.pathname,
         gamer_tag: state.settings.gamer_tag ?? null,
-        click_behavior: state.settings.click_behavior ?? "single",
+        click_behavior: state.settings.click_behavior ?? "double",
       };
       void saveSettings(updated).catch(() => {
         // Silently ignore -- route restore is not critical.
